.vscode/requirements.txt
opencv-python
numpy
pillow
pystray
pyinstaller